model06: clean up code and use avg positions for uid and qid for test
In this model, we use average positions for uid and qid for test set.
However, in test set, there are many qid and uid which do not exist
in train set. So, I just use mean of avg positions for uid and qid
in this case.

It earns 85.85977 by Lasso and 87.05909 by Ridge. I think we need to
check our implementation at this moment especially for feature
engineering.